+++ CHANGELOG.md
@@ -14,6 +14,8 @@ The format is based on [Keep a Changelog] and this project adheres to [Semantic 
 - Set edition to 2024
 - Removed `Error::DeleteDirAsFile` (breaking change)
 - Renamed `delete_file_in_dir` to `delete_entry_in_dir` because it can now delete empty directories (breaking change)
+- Added `open_long_name_file_in_dir` API, to open files using their Long File Name
+- Updated directory iterator callback, to allow bailing out early (breaking change)
 
 ## [Version 0.9.0] - 2025-06-08